requirements.txt
Flask
tensorflow==2.19.0
numpy
opencv-python-headless
python-dotenv
gdown
gunicorn
